assertion for parameter types

+++ SomaticSiMu.py
@@ -694,11 +694,11 @@ def somatic_sim(cancer_type, reading_frame, std_outlier, sequence_abs_path, slic
         print("power argument must be int type")
         sys.exit()
         
-    if isinstance(syn_rate, int) == False:
+    if isinstance(syn_rate, float) == False:
         print("syn_rate argument must be int type")
         sys.exit()
         
-    if isinstance(non_syn_rate, int) == False:
+    if isinstance(non_syn_rate, float) == False:
         print("non_syn_rate argument must be int type")
         sys.exit()
         
@@ -1589,9 +1589,9 @@ def main():
             cancer_type = item
             reading_frame = int(args.reading_frame)
             std_outlier = int(args.std)
-            power = args.power
-            syn_rate = args.syn_rate
-            non_syn_rate = args.non_syn_rate
+            power = int(args.power)
+            syn_rate = float(args.syn_rate)
+            non_syn_rate = float(args.non_syn_rate)
                         
             func = partial(somatic_sim, cancer_type, reading_frame, std_outlier, sequence_abs_path, slice_start, slice_end, power, syn_rate, non_syn_rate, sample_seq, sample_index_dict, k1mer_count_map, k2mer_count_map, k3mer_count_map, k4mer_count_map, k5mer_count_map, k6mer_count_map)
